+++ src/test/java/com/tpokora/projects/user/utils/UserTestUtils.java
@@ -1,11 +1,9 @@
 package com.tpokora.projects.user.utils;
 
-import com.tpokora.projects.common.model.TableDetails;
 import com.tpokora.projects.common.utils.TestUtils;
 import com.tpokora.projects.user.model.User;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 /**
@@ -17,7 +15,7 @@ public class UserTestUtils extends TestUtils {
         List<User> usersList = new ArrayList<User>();
         if (amount > 0) {
             for (int i = 1; i <= amount; i++) {
-                usersList.add(new User(Long.valueOf(i), "user" + i, "pass", "email@email.com", "ROLE_USER"));
+                usersList.add(new User(Integer.valueOf(i), "user" + i, "pass", "email@email.com", "ROLE_USER"));
             }
         }
